+++ src/alx/Enemy.h
@@ -66,10 +66,10 @@ struct Enemy {
 
     void draw(std::vector<uint32_t>& screen_buffer, float alpha) const {
         Draw::rect(
-            static_cast<int>(x),
-            static_cast<int>(y),
-            static_cast<int>(width),
-            static_cast<int>(height),
+            x,
+            y,
+            width,
+            height,
             color,
             true, // fill
             1,    // thickness
